--- examples/taint/taint_direct.h
@@ -1,8 +1,0 @@
-#define RESULT_ASSERT_NESTED_CONDITIONAL 0
-#define RESULT_ASSERT_NESTED_FUN 0
-#define RESULT_TEST_WHILE 0
-#define RESULT_ASSERT_TEST_CONDITIONAL 0
-#define RESULT_CALL_RET 3
-#define RESULT_ASSERT_TEST_CONDITIONAL_REGIONS 0
-#define RESULT_ASSERT_TEST_BR 0
-
