--- examples/taint/taint_regions.h
@@ -1,6 +1,0 @@
-#define RESULT_ASSERT_NESTED_CONDITIONAL 1
-#define RESULT_ASSERT_NESTED_FUN 1
-#define RESULT_TEST_WHILE 1
-#define RESULT_ASSERT_TEST_CONDITIONAL 1
-#define RESULT_ASSERT_TEST_CONDITIONAL_REGIONS 0
-#define RESULT_ASSERT_TEST_BR 0
