+++ lib/Core/ExecutionState.cpp
@@ -152,7 +152,7 @@ TaintSet ExecutionState::getPCTaint() {
 //current Program counter taint setter
 void ExecutionState::setPCTaint(TaintSet new_taint) {
     taint = new_taint;
-    //std::cerr << "PC TAINTED! " << taint <<std::endl;
+    KLEE_DEBUG(llvm::errs() << "PC TAINTED! " << taint <<std::endl);
 }
 
 //get the depth of the SESE region stack
@@ -163,14 +163,14 @@ int ExecutionState::getRegionDepth() {
 //called when entering a new SESE region 
 void ExecutionState::enterRegion() {
     stack.back().regionStack.push(taint);
-    //std::cerr << "REGION CHANGED! PUSH!" << taint <<std::endl;
+    KLEE_DEBUG(llvm::errs() << "REGION CHANGED! PUSH!" << taint <<std::endl;
 }
 
 //called when leaving a SESE region 
 void ExecutionState::leaveRegion() {
     taint = stack.back().regionStack.top();
     stack.back().regionStack.pop();
-    //std::cerr << "REGION CHANGED! POP!" << taint <<std::endl;
+    
 }
