+++ lib/Module/KModule.cpp
@@ -14,6 +14,7 @@
 #include "klee/Internal/Module/KModule.h"
 
 #include "Passes.h"
+#include "llvm/Analysis/Passes.h"
 
 #include "klee/Config/Version.h"
 #include "klee/Interpreter.h"
@@ -57,8 +58,10 @@
 #include "llvm/Transforms/Scalar.h"
 
 #include <llvm/Transforms/Utils/Cloning.h>
-#include <llvm/Analysis/RegionInfo.h>
 #include <sstream>
+
+#include <llvm/Analysis/RegionInfo.h>
+#include <llvm/InitializePasses.h>
 
 using namespace llvm;
 using namespace klee;
@@ -204,38 +207,22 @@ static void injectStaticConstructorsAndDestructors(Module *m) {
   }
 }
 //////////////////////////////////////////////////////////////// Begin auxiliary taint region Function pass
-#include "llvm/InitializePasses.h"
 // TaintRegionsInfo - The first implementation, without getAnalysisUsage.
-struct TaintRegionsInfoPass : public  ModulePass {
+struct TaintRegionsInfoPass : public RegionInfo {
   static char ID; // Pass identification, replacement for typeid
-  TaintRegionsInfoPass(std::map<llvm::BasicBlock *, int > &regions) : ModulePass(ID), br(regions) {  }
-  Module *M;
-  std::map<BasicBlock *, int> &br; 
+  std::map<BasicBlock *, int> *br; 
+  TaintRegionsInfoPass(std::map<llvm::BasicBlock *, int > *regions) : br(regions) { llvm::errs() <<"WTF!@"; }
 
-    virtual bool runOnModule(Module &M) {
-      llvm::RegionInfo &RI = this->getAnalysis<llvm::RegionInfo>();
-
-
-
-
-      for (Module::iterator F = M.begin(), fe = M.end(); F != fe; ++F)
-
-      for (Function::iterator bb = F->begin(), e = F->end(); bb != e; ++bb)
-        br.insert(std::make_pair(bb, RI.getRegionFor(bb)->getDepth()));
-//((std::string) F.getName()) +": "+ RI.getRegionFor(bb)->getNameStr()
-      return false;
-    }
-    // We don't modify the program, so we preserve all analyses
-    virtual void getAnalysisUsage(AnalysisUsage &AU) const {
-       AU.addRequired<RegionInfo>();
-       AU.setPreservesAll();
-    }
-
-
+  virtual bool runOnFunction(Function &F){
+      bool result = RegionInfo::runOnFunction(F);
+      for (Function::iterator bb = F.begin(), e = F.end(); bb != e; ++bb){
+          br->insert(std::make_pair(bb, this->getRegionFor(bb)->getDepth()));
+	  }
+      return result;
+  }
+  virtual const char* getPassName() const { return "TaintRegionInfo"; }
 };
-char TaintRegionsInfoPass::ID = 0;
-//INITIALIZE_PASS(TaintRegionsInfoPass, "TaintRegionsInfo", "TaintRegionsInfo World Pass", false, true);
-//////////////////////////////////////////////////////////////// End aucxiliary taint region Function pass
+//////////////////////////////////////////////////////////////// End auxiliary taint region Function pass
 
 #if LLVM_VERSION_CODE < LLVM_VERSION(3, 3)
 static void forceImport(Module *m, const char *name, LLVM_TYPE_Q Type *retType,
@@ -345,23 +332,13 @@ void KModule::prepare(const Interpreter::ModuleOptions &opts,
   if (opts.Optimize)
     Optimize(module);
 
-  //
+  //Computing region info
   if (opts.CalculateRegions){
-    llvm::errs() << "KLEE: Computing region info\n";
-    PassManager Passes;
-    TaintRegionsInfoPass *regionsInfo = new TaintRegionsInfoPass(this->regions);
-    Passes.add((FunctionPass *)regionsInfo);
-    Passes.run(*module);
-    //this->regionsInfo contains SESE region iformation for the module 
-    /* FIX DEBUG TODO 
-    printf("CALCULATED REGIONS!!\n");
-    for ( std::map<BasicBlock *, std::string >::iterator it = this->regions.begin();
-                    it!= this->regions.end();
-                    it++)
-            printf("%p -> %s\n",it->first, it->second.c_str());
-    */
-    }
-
+    PassManager pm;
+    pm.add(new TaintRegionsInfoPass(&this->regions));
+    pm.run(*module);
+    //this->regions contains SESE region iformation for the module 
+  }
 #if LLVM_VERSION_CODE < LLVM_VERSION(3, 3)
   // Force importing functions required by intrinsic lowering. Kind of
   // unfortunate clutter when we don't need them but we won't know
